==> 1_Parse_CUAD_V1.ipynb <==
import re
import os
import itertools
import numpy as np
import pandas as pd
import nltk
from tqdm import tqdm

PATH_FOLDER_TXT = "./data/CUAD_v1/full_contract_txt/"
PATH_LABELS = './data/CUAD_v1/master_clauses.csv'

PATH_OUTPUT = './data/glaw/glaw_data.csv'

text_file_names = []
for (dirpath, dirnames, filenames) in os.walk(PATH_FOLDER_TXT):
    text_file_names.extend(filenames)

text_file_names.sort()

df_labels = pd.read_csv(PATH_LABELS)

df_main = df_labels[['Filename', 'Governing Law', 'Governing Law-Answer']].copy()
df_main.sort_values('Filename', axis=0, inplace=True, ignore_index=True)

# Bring in the list of the .txt filenames
df_main.insert(loc=1, column='text_file_name', value=text_file_names)
df_main.info()

df_main["Governing Law"] = df_main["Governing Law"].apply(eval)
gw_empty_mask = df_main['Governing Law'].apply(lambda it: len(it) == 0)
df_main[gw_empty_mask] = np.nan
df_main['Governing Law-Answer'].replace('[]', np.NaN, inplace=True)
df_main['Governing Law-Answer'].replace('[* * *]', np.NaN, inplace=True)
df_main.dropna(axis=0, how='any', inplace=True)
df_main.info()

def pre_process_doc_common(text):
    # Simple replacement for "#"
    text = text.replace("#", " ")

    # Simple replacement for "\n"
    text = text.replace("\n", " ")

    # Simple replacement for "\xa0"
    text = text.replace("\xa0", " ")

    # Simple replacement for "\x0c"
    text = text.replace("\x0c", " ")

    # insert spaces around to get this as token if future
    text = text.replace('<omitted>', ' <omitted> ')

    # Get rid of multiple dots
    regex = "\\ \\.\\ "
    subst = "."
    text = re.sub(regex, subst, text, 0)

    # Get rid of underscores
    regex = "_"
    subst = " "
    text = re.sub(regex, subst, text, 0)

    # Get rid of multiple dashes
    regex = "--+"
    subst = " "
    text = re.sub(regex, subst, text, 0)

    # Get rid of multiple stars
    regex = "\\*+"
    subst = "*"
    text = re.sub(regex, subst, text, 0)

    # Get rid of multiple whitespace
    regex = " +"
    subst = " "
    text = re.sub(regex, subst, text, 0)

    # Strip leading and trailing whitespace
    text = text.strip()

    return text


# Function to take in the file list, read each file, clean the text and return all agreements in a list
def get_text_data_cleared(file_path_list, print_text=False, clean_text=True):
    text_list = []
    for file_path in tqdm(file_path_list):
        with open(file_path, "r", encoding="utf8") as agreement:
            text = agreement.read()
            if print_text:
                print("Text before cleaning: \n", text)

        if clean_text:
            text = pre_process_doc_common(text)

        if print_text:
            print("Text after cleaning: \n", text)

        text_list.append(text)

    return text_list

print('b\xa0a')

file_list = [os.path.join(PATH_FOLDER_TXT, row['text_file_name']) for index, row in df_main.iterrows()]
text_cleared_list = get_text_data_cleared(file_list, print_text=False, clean_text=True)
df_main.insert(loc=2, column='text', value=text_cleared_list)
df_main.head()

def find_labels(labels_to_find, txt, key):
    result = []
    for lbl in labels_to_find:
        lbl_cleared = pre_process_doc_common(lbl)
        if len(lbl_cleared):
            matches = list(re.finditer(re.escape(lbl_cleared.lower()), txt.lower()))
            if len(matches) == 0:
                print(f'Failed to find {lbl_cleared}')
            for m in matches:
                s = m.start()
                e = m.end()
                result.append((s, e, key))

    return result

# Go through each label and find the label in the text, ensure label is pre-processed same as text.
# If labels don't match, append to a separate file to check.
def sort_df(df):
    djson = {}
    djson_inspect = {}
    for index, row in tqdm(df.iterrows()):
        labels = list()
        text = row['text']

        #GOVERNING_LAW
        g_laws = row['Governing Law']
        if g_laws is not None and len(g_laws) > 0:
            g_law_labels = find_labels(g_laws, text, 'GOVERNING_LAW')
            labels.extend(g_law_labels)

        # Check for incongruous finds, add to inspect file
        labels = tuple(labels)
        if len(labels) > 0:
            djson[index] = labels
        else:
            djson_inspect[index] = labels

    return djson, djson_inspect

djson, djson_inspect = sort_df(df_main)
djson_inspect

df_main['Governing Law Occurrences'] = djson.values()
df_main.head()

jurisdictions_preprocessed = [pre_process_doc_common(location) for location in df_main['Governing Law-Answer']]
pattern_split_jurisdiction = r'; ?|, ?'
jurisdiction_splitted = [re.split(pattern_split_jurisdiction, location) for location in jurisdictions_preprocessed]
jurisdictions = list(itertools.chain(*jurisdiction_splitted))
df_main['Governing Law-Answer Splitted'] = jurisdiction_splitted
df_main.head()

from data.data import states

glaw_search_anchors = set(jurisdictions + states)
glaw_search_anchors.remove('Washington')  # there is more sophisticated pattern
glaw_search_anchors.add('PRC')  # alias for People's Republic of China
glaw_search_anchors.add('English')  # this form was missed
glaw_search_anchors.add('Italian')  # this form was missed
glaw_search_anchors.add('Swiss')    # this form was missed
print(len(glaw_search_anchors))
glaw_search_anchors

def has_intersection(first_span, second_span):
    if first_span[0] <= second_span[0] < first_span[1] or second_span[0] <= first_span[0] < second_span[1]:
        return True

    return False

# better approach would be to add to candidate_text_list true_occurrences and all occurrence candidates found that don't intersect with any of true_occurrences
sentence_tokenizer = nltk.data.load('tokenizers/punkt/english.pickle')
pattern_glaw_search = '|'.join(glaw_search_anchors)
regex_glaw_search = re.compile(pattern_glaw_search, flags=re.IGNORECASE)
candidates_list = []
labels_list = []
for idx, row in df_main.iloc[0:].iterrows():
    text = row['text']
    true_occurrences = row['Governing Law Occurrences']
    true_anchors = row['Governing Law-Answer Splitted']
    # crutches
    if "People's Republic of China" in true_anchors:
        true_anchors.append('PRC')
    if 'England' in true_anchors:
        true_anchors.append('English')
    if 'Italy' in true_anchors:
        true_anchors.append('Italian')
    if 'Switzerland' in true_anchors:
        true_anchors.append('Swiss')

    sentences = list(sentence_tokenizer.span_tokenize(text))
    sentences_used = [False] * len(sentences)

    candidate_text_list = []
    candidate_label_list = []
    for match in regex_glaw_search.finditer(text):
        span = match.start(), match.end()
        for s_idx, s in enumerate(sentences):
            if has_intersection(span, s) and not sentences_used[s_idx]:
                sentences_used[s_idx] = True
                candidate_text = text[s[0]: s[1]]
                if s_idx > 0:
                    prev_s = sentences[s_idx - 1]
                    candidate_text = text[prev_s[0]: prev_s[1]] + ' ' + candidate_text

                is_true = False
                for true_s in true_occurrences:
                    if has_intersection(s, true_s):
                        for true_a in true_anchors:
                            if re.search(true_a, text[s[0]: s[1]], flags=re.IGNORECASE):
                                is_true = True
                                break

                candidate_text_list.append(candidate_text)
                candidate_label_list.append(int(is_true))
                break

    # workarounds
    if idx == 252:
        candidate_label_list[3] = 0
    elif idx == 326:
        del candidate_text_list[3]
        del candidate_label_list[3]
    elif idx == 382:
        candidate_label_list[6] = 0
    elif idx == 448:
        candidate_label_list[2] = 0
    doc_exception_idx_list = [145]  # these docs checked, such behavior is correct
    if len(true_occurrences) != sum(candidate_label_list) and idx not in doc_exception_idx_list:
        for s in true_occurrences:
            print(text[s[0]: s[1]])
            print()
        print(true_occurrences)
        print(candidate_text_list[2:])
        print(candidate_label_list)
        raise Exception(f'Error at document {idx}')

    candidates_list.append(candidate_text_list)
    labels_list.append(candidate_label_list)

df_main['candidates'] = candidates_list
df_main['labels'] = labels_list
df_main.head()

df_main.to_csv(PATH_OUTPUT, columns=['candidates', 'labels'])
