==> 2_Vectorize_GPComments.ipynb <==
import os
import glob

from tqdm import tqdm
import numpy as np
import pandas as pd
from joblib import dump
from scipy.sparse import save_npz
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS

from utils.lemma_tokenizer import LemmaTokenizer

PATH_DATA = './data/GPCommentsRaw'

search_path = os.path.join(PATH_DATA, '*.json')
data_files_list = glob.glob(search_path)
data_frame_list = []
for path in tqdm(data_files_list):
    df = pd.read_json(path, orient='records')
    data_frame_list.append(df[['text', 'score']])

df = pd.concat(data_frame_list, axis=0)
df.head()

print(df['text'].isna().sum())

df.dropna(axis=0, how='all', subset='text', inplace=True)
print(df['text'].isna().sum())

df['labels'] = (df['score'] >= 3).astype(np.int32).values
df.tail()

data_version = 0

pattern_token = r'(?u)(\b[a-z]{2,}\b|[\u263a-\U0001f645]|!|\?)'
tokenizer = LemmaTokenizer(pattern_token)
stop_words_lematized = [tokenizer.lemmatizer.lemmatize(word) for word in ENGLISH_STOP_WORDS]
stop_words_lematized.append('shall')
vectorizer = CountVectorizer(
                            tokenizer=tokenizer,
                            strip_accents='unicode',
                            lowercase=True,
                            stop_words=stop_words_lematized,
                            ngram_range=(1, 2),
                            min_df=10_000,
                            binary=True
)

text_vectorized = vectorizer.fit_transform(df['text'])

mask = text_vectorized.getnnz(axis=1) > 1
text_vectorized_cleared = text_vectorized[mask]
df_cleared = df[mask]
df_cleared.shape

text_vectorized.shape, text_vectorized_cleared.shape

df_cleared.to_csv('./data/gp_comments/targets.csv', columns=['score', 'labels'])

vectorizer_params = {
    'vocabulary_': vectorizer.vocabulary_,
}
dump(vectorizer_params, f'./data/gp_comments/data_headers_{data_version}.sav')

save_npz(f'./data/gp_comments/data_vectorized_{data_version}.npz', text_vectorized_cleared)
