==> 3_Classify_GPComments.ipynb <==
import pandas as pd
import numpy as np
from scipy.sparse import load_npz
from joblib import load
from sklearn.naive_bayes import MultinomialNB
from scipy.special import logsumexp

data_version = 0

data = load_npz(f'./data/gp_comments/data_vectorized_{data_version}.npz')
labels = pd.read_csv('./data/gp_comments/targets.csv')['labels']
print(data.shape[0], len(labels))

model = MultinomialNB()

batch_size = 10_000
for start in range(0, data.shape[0], batch_size):
    print(start)
    x_batch, y_batch = data[start: start+batch_size], labels[start: start+batch_size]
    model.partial_fit(x_batch, y_batch, classes=[0, 1])

# predictions = model.predict(data)

# from sklearn.metrics import recall_score, precision_score, f1_score
# print(f'recall: {recall_score(labels, predictions)}')
# print(f'precision: {precision_score(labels, predictions)}')
# print(f'f1 score: {f1_score(labels, predictions)}')

log_likelihood = model.feature_log_prob_ + model.class_log_prior_.reshape((-1, 1))
log_prob_x = logsumexp(log_likelihood, axis=0)
log_proba = log_likelihood - np.atleast_2d(log_prob_x)
positive_class_probabilities = np.exp(log_proba)[1]

headers = load(f'./data/gp_comments/data_headers_{data_version}.sav')['vocabulary_']
headers_ordered = sorted(headers.keys(), key=lambda key: headers[key])

df_result = pd.DataFrame({
    'word': headers_ordered,
    'probability': positive_class_probabilities,
})
df_result.head(50)

df_result.to_csv(f'./data/gp_comments/word_probabilities_{data_version}.csv')

df_result_sorted = df_result.sort_values('probability', ascending=False, ignore_index=True)
df_result_sorted.head(50)
