==> 4_Vectorize_GLaw_data.ipynb <==
from itertools import chain

import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS
from joblib import dump

from utils.lemma_tokenizer import LemmaTokenizer

df = pd.read_csv('./data/glaw/glaw_data.csv', index_col=0)
df.head()

df['candidates'] = df['candidates'].apply(eval)
df['labels'] = df['labels'].apply(eval)
print(type(df.iloc[0]['candidates']))
print(type(df.iloc[0]['labels']))

data_version = 1

pattern_token = r'(?u)(\b[a-z]{2,}\b)'
tokenizer = LemmaTokenizer(pattern_token)
stop_words_lematized = [tokenizer.lemmatizer.lemmatize(word) for word in ENGLISH_STOP_WORDS]
stop_words_lematized.append('shall')
vectorizer = CountVectorizer(
    tokenizer=tokenizer,
    strip_accents='unicode',
    lowercase=True,
    stop_words=stop_words_lematized,
    ngram_range=(1, 2),
    min_df=10,
    binary=True
)

all_text_list = chain.from_iterable(df['candidates'])
vectorizer.fit(all_text_list)
len(vectorizer.vocabulary_)

multi_candidate_mask = df['labels'].apply(lambda l: 0 in l).to_numpy()
df = df[multi_candidate_mask]
df.shape

data_vectorized = []
for idx, row in df.iterrows():
    doc_candidate_list = row['candidates']
    labels = np.asarray(row['labels'])

    text_vectorized = vectorizer.transform(doc_candidate_list)
    mask = text_vectorized.getnnz(axis=1) > 1
    text_vectorized = text_vectorized[mask]
    labels = labels[mask]

    if 1 in labels:
        # only if at least 1 positive example present
        row.at['labels'] = list(labels)
        data_vectorized.append(text_vectorized)
len(data_vectorized)

vectorizer_params = {
    'vocabulary_': vectorizer.vocabulary_,
}
dump(vectorizer_params, f'./data/glaw/data_headers_{data_version}.sav')

dump(data_vectorized, f'./data/glaw/data_vectorized_{data_version}.sav')

df.to_csv(f'./data/glaw/labels_{data_version}.csv', columns=['labels'])
